Fixed typo (float -> double)

+++ include/Triangulate/vector.h
@@ -29,7 +29,7 @@ double dot(const Vector& a, const Vector& b);
 double det(const Vector& a, const Vector& b);
 double angleBetween(const Vector& a, const Vector& b);
 
-bool equal(const float& a, const float& b);
+bool equal(const double& a, const double& b);
 bool cwLess(const Vector& a, const Vector& b, const Vector& c);
 bool operator==(const Vector& a, const Vector& b);
 bool operator!=(const Vector& a, const Vector& b);
